School_Project/requirements.txt
amqp==5.1.1
asgiref==3.5.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Django==4.1.1
django-bootstrap-v5==1.0.11
django-celery-beat==2.4.0
django-debug-toolbar==3.7.0
django-mysql==3.8.1
django-timezone-field==5.0
djangorestframework==3.14.0
Faker==15.0.0
kombu==5.2.4
lorem-text==2.1
mysqlclient==2.1.1
prompt-toolkit==3.0.32
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
timedelta==2020.12.3
tzdata==2022.4
vine==5.0.0
wcwidth==0.2.5
